backend/requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
openai-whisper
